feat!: Rename material 3 specific classes
