+++ components/src/androidTest/java/uk/gov/android/ui/components/ButtonsTest.kt
@@ -13,6 +13,9 @@ import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.junit.runners.Parameterized
+import uk.gov.android.ui.components.buttons.ButtonParameters
+import uk.gov.android.ui.components.buttons.ButtonProvider
+import uk.gov.android.ui.components.buttons.GdsButton
 import uk.gov.android.ui.theme.GdsTheme
 import uk.gov.android.ui.theme.minimumTouchTarget
