+++ pages/src/main/java/uk/gov/android/ui/pages/calculator/InformationScreen.kt
@@ -19,12 +19,12 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.tooling.preview.PreviewParameterProvider
 import androidx.compose.ui.unit.Dp
-import uk.gov.android.ui.components.ButtonParameters
-import uk.gov.android.ui.components.ButtonType
-import uk.gov.android.ui.components.GdsButton
 import uk.gov.android.ui.components.GdsHeading
 import uk.gov.android.ui.components.HeadingParameters
 import uk.gov.android.ui.components.HeadingSize
+import uk.gov.android.ui.components.buttons.ButtonParameters
+import uk.gov.android.ui.components.buttons.ButtonType
+import uk.gov.android.ui.components.buttons.GdsButton
 import uk.gov.android.ui.components.content.BulletListParameters
 import uk.gov.android.ui.components.content.ContentParameters
 import uk.gov.android.ui.components.content.GdsBulletList
